Fix infinite loop on comment line break in JS related modes
Using tree-sitter to customize the behavior of comment newlines.
Previously used to cause lisp max-depth issues when hitting RET with the
point at the very beginning of the comment. Hopefully this might be
faster than the smartparens default too? though I'm not really sure.

Also override js2-line-break to not call `newline-and-indent` directly